pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "pr-guardian"
version = "0.1.0"
description = "PR Guardian: pull request quality gate and automation helper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "click>=8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "W", "F", "I", "B", "UP"]

[tool.mypy]
strict = true
python_version = "3.11"
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
